chore: Align Feature Folder Structure
- add feature folder for Navigation and (Side-)/(Bottom-)Sheet
