+++ app/src/main/kotlin/org/eclipse/kuksa/companion/feature/navigation/view/HorizontalNavigationView.kt
@@ -17,7 +17,7 @@
  *
  */
 
-package org.eclipse.kuksa.companion.feature.home.view.navigation
+package org.eclipse.kuksa.companion.feature.navigation.view
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -34,6 +34,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import org.eclipse.kuksa.companion.feature.navigation.NavigationPage
+import org.eclipse.kuksa.companion.feature.navigation.viewmodel.NavigationViewModel
 
 @Composable
 fun HorizontalNavigationView(
